+++ demos/src/bin/platformer_2d.rs
@@ -210,6 +210,7 @@ fn setup_player(mut commands: Commands) {
         dash: Default::default(),
         one_way_platforms_min_proximity: 1.0,
         falling_through: FallingThroughControlScheme::SingleFall,
+        knockback: Default::default(),
     });
 
     // An entity's Tnua behavior can be toggled individually with this component, if inserted.