+++ demos/src/character_animating_systems/platformer_animating_systems.rs
@@ -1,6 +1,7 @@
-use bevy::ecs::query::QueryData;
 use bevy::prelude::*;
-use bevy_tnua::builtins::{TnuaBuiltinCrouch, TnuaBuiltinDash, TnuaBuiltinJumpState};
+use bevy_tnua::builtins::{
+    TnuaBuiltinCrouch, TnuaBuiltinDash, TnuaBuiltinJumpState, TnuaBuiltinKnockback,
+};
 use bevy_tnua::math::Float;
 use bevy_tnua::prelude::*;
 use bevy_tnua::{TnuaAnimatingState, TnuaAnimatingStateDirective};
@@ -16,54 +17,7 @@ pub enum AnimationState {
     Crouching,
     Crawling(Float),
     Dashing,
-    KnockedBack(Dir3),
-}
-
-impl AnimationState {
-    fn force_forward(&self) -> Option<Dir3> {
-        match self {
-            AnimationState::Standing => None,
-            AnimationState::Running(_) => None,
-            AnimationState::Jumping => None,
-            AnimationState::Falling => None,
-            AnimationState::Crouching => None,
-            AnimationState::Crawling(_) => None,
-            AnimationState::Dashing => None,
-            AnimationState::KnockedBack(direction) => Some(-*direction),
-        }
-    }
-}
-
-#[derive(QueryData)]
-#[query_data(mutable)]
-pub struct ForwardForcing {
-    pub transform: &'static mut Transform,
-    #[cfg(feature = "rapier3d")]
-    rapier3d_locked_axes: &'static mut bevy_rapier3d::prelude::LockedAxes,
-    #[cfg(feature = "avian3d")]
-    avian3d_locked_axes: &'static mut avian3d::prelude::LockedAxes,
-}
-
-impl ForwardForcingItem<'_> {
-    fn lock_rotation(&mut self) {
-        #[cfg(feature = "rapier3d")]
-        self.rapier3d_locked_axes
-            .insert(bevy_rapier3d::prelude::LockedAxes::ROTATION_LOCKED_Y);
-        #[cfg(feature = "avian3d")]
-        {
-            *self.avian3d_locked_axes = self.avian3d_locked_axes.lock_rotation_y();
-        }
-    }
-
-    fn unlock_rotation(&mut self) {
-        #[cfg(feature = "rapier3d")]
-        self.rapier3d_locked_axes
-            .remove(bevy_rapier3d::prelude::LockedAxes::ROTATION_LOCKED_Y);
-        #[cfg(feature = "avian3d")]
-        {
-            *self.avian3d_locked_axes = self.avian3d_locked_axes.unlock_rotation_y();
-        }
-    }
+    KnockedBack,
 }
 
 #[allow(clippy::unnecessary_cast)]
@@ -80,13 +34,10 @@ pub fn animate_platformer_character(
         // for deciding which animation to play.
         &TnuaController,
         &AnimationsHandler,
-        ForwardForcing,
     )>,
     mut animation_players_query: Query<&mut AnimationPlayer>,
 ) {
-    for (mut animating_state, controller, handler, mut forward_forcing) in
-        animations_handlers_query.iter_mut()
-    {
+    for (mut animating_state, controller, handler) in animations_handlers_query.iter_mut() {
         let Ok(mut player) = animation_players_query.get_mut(handler.player_entity) else {
             continue;
         };
@@ -138,6 +89,7 @@ pub fn animate_platformer_character(
                 // For the dash, we don't need the internal state of the dash action to determine
                 // the action - so there is no need to downcast.
                 Some(TnuaBuiltinDash::NAME) => AnimationState::Dashing,
+                Some(TnuaBuiltinKnockback::NAME) => AnimationState::KnockedBack,
                 Some(other) => panic!("Unknown action {other}"),
                 None => {
                     // If there is no action going on, we'll base the animation on the state of the
@@ -146,9 +98,7 @@ pub fn animate_platformer_character(
                     else {
                         continue;
                     };
-                    if let Some(pushover_direction) = basis_state.pushover() {
-                        AnimationState::KnockedBack(pushover_direction)
-                    } else if basis_state.standing_on_entity().is_none() {
+                    if basis_state.standing_on_entity().is_none() {
                         AnimationState::Falling
                     } else {
                         let speed = basis_state.running_velocity.length();
@@ -226,24 +176,13 @@ pub fn animate_platformer_character(
                     AnimationState::Dashing => {
                         player.start(handler.animations["Dashing"]).set_speed(10.0);
                     }
-                    AnimationState::KnockedBack(_) => {
+                    AnimationState::KnockedBack => {
                         player
                             .start(handler.animations["KnockedBack"])
                             .set_speed(1.0);
                     }
                 }
             }
-        }
-
-        // The knockback animation needs the character to be in a specific direciton. We want the
-        // character to face that way immediately (a gradual turn would look weird here) and we
-        // also want to lock the rotation so that the physics engine won't be able to slightly
-        // change it between frames (which Tnua will try to do)
-        if let Some(forward) = animating_state.get().and_then(|s| s.force_forward()) {
-            forward_forcing.lock_rotation();
-            forward_forcing.transform.look_to(forward, Dir3::Y);
-        } else {
-            forward_forcing.unlock_rotation();
         }
     }
 }