+++ src/util/mod.rs
@@ -1,5 +1,3 @@
-pub mod boundary;
-
 use bevy::prelude::*;
 use bevy_tnua_physics_integration_layer::math::{
     AdjustPrecision, Float, Quaternion, Vector2, Vector3,